pyproject.toml
[project]
name = "pyforth"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "smeyn", email = "stephanmeyn@gmail.com" }
]
dependencies = [
    "pytest>=8.3.4",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pyforth"]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]
pythonpath = "src"

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "pip>=24.3.1",
    "rio-ui>=0.10.9",
]

 